requirements.txt
ultralytics
opencv-python-headless
pyyaml
argparse
streamlit
psutil
onnxruntime
pyinstaller